Wire window size propagation
Events on window size need to be propagated to the WindowPod so they
can be available during layout.

There are other ways to do this. Instead of being in WindowPod, maybe
the size info should be in WindowState (where prev_paint_time is, for
similar reasons), then passed down through LayoutCtxRoot.

But this way, at least the RootWidget gets notifications of window size
changing, which seems potentially useful.